gedit: port GeditFileChooserDialogGtk to GtkFileChooserNative
The intent is to allow GEdit to make use of portals when running under
confinement, removing the need to give it wide file system access.

The encodings and newline combo boxes are converted to use the newer
"choice" API, which handles everything except adding new encodings to
the list.

The file filters code also remains mostly unchanged.  The use of a
custom file filter will need some changes too.