docs: gedit-development-getting-started: copy from the wiki
Copied from:
https://wiki.gnome.org/Apps/Gedit/DevGettingStarted?action=raw&rev=12

(At revision 12, not the last revision because some important infos have
been removed. I'm the author of that wiki page, and revision 12 is my
last edit to that page. I will see if I can take some paragraphs from
the last revision, the new edits have been made by a newcomer, so it's a
good stress test for that wiki page, to see if a newcomer succeeds to
get started with gedit development. So basically I will take the best of
both revisions.)