flatpak: temporarily disable building the gedit-plugins repo
Just to be sure that the GitLab CI can now build the flatpak, now that
it depends on Tepl.

With gedit-plugins it cannot find the zeitgeist dependency. It was not
bundled in the gedit flatpak manifest before, so I wonder how it worked
before, probably because it was part of the runtime and it is no longer
the case? Do we really need zeitgeist? Are people actually using it?