Bump API version to 3.32
There was an API break in
commit 742a7c046a2336b219a22773343c80e38b78339b

Using a new major version of a library, with that library exposed in the
gedit plugin API, it's an API break for gedit plugins. In this case
GtkSourceView 3 -> GtkSourceView 4.

`git tag --contains 742a7c046a2336b219a22773343c80e38b78339b` returns
that the 3.32 version is the first stable version containing the API
break.