gedit-dirs: remove get_user_cache_dir() (no longer used)
It was used for the metadata manager, but it's not really the
appropriate place. Metadata is not a cache, it's user data. Without the
file metadata, the users lose some information (and the next time that
the file is loaded, it relies on automatic/default behavior).

tepl_metadata_manager_init() recommends to use g_get_user_data_dir().