Bump API version 3.36
The API version was bumped to 3.32 recently. Each time that there is an
API version bump, all the deprecated API is removed, normally (during
the same development cycle as the API version bump).

But here, 3.32 has already been released. It's currently the 3.35
development cycle. And the deprecated API has been removed now, not for
gedit 3.32. On top of that gedit 3.32 had a different API version, so it
complicates things.

So to make things clearer, this commit bumps the API version to 3.36.