main: fix program name in `gedit --help` output
The output was:
```
$ gedit --help
Usage:
  org.gnome.gedit [OPTION…] [FILE…] [+LINE[:COLUMN]]
[...]
```

There was a comment that links to a blog post. After reading the blog
post, I don't think gedit needs it, because everything has the
'org.gnome.gedit' ID/basename: the *.desktop file, the icon, the
GApplication ID.

If calling g_set_prgname() is really needed, then it needs to be done at
a different place, to not break `gedit --help` for normal usage, and not
break all the calls to g_get_prgname().