Revert "gedit-commands-file.c: Only save if file has been modified"
This reverts commit deb2af02a9dad8fe2d1cb8689c55a9ac54a8553f.

It's important to be able to save a document in all circumstances.

Opening a text file in gedit and then directly save it is like the
equivalent of the `touch` Unix command, but with a GUI.

Also, a file can be modified with another program while it is opened in
gedit. In that case it is useful to be able to save the file again in
gedit. gedit tries to detect when a file is externally modified, but it
does it only for local files.

Another reason is if gedit bugs, if the GeditTab/GeditDocument state
(wrt whether it needs saving or not) is buggy, then the user will be
happy to be able to save the content without the need to copy/paste the
text into another app.

And there are probably other reasons why disabling a "save" feature can
be dangerous. On Unix when we execute a command, the system listens to
the user and executes the command, regardless of whether it's useful or
not.