App: do not call gtk_window_set_role() (specific to X11)
To try to get rid of X11-specific things in gedit.

This also permits to get rid of one use of g_get_current_time() and
GTimeVal (deprecated, not 2038-year-safe).

gtk_window_get_role() is not called in gedit. And with the unique string
for the window role, the window manager was never able to re-use its
knowledge about gedit windows.