build: require libsoup dep only on Windows for the checkupdate plugin
libsoup is used nowhere else in this git repo, it's used only by the
checkupdate plugin. The checkupdate plugin needs access to the internet,
but it's the only plugin requiring that access. So when sandboxing
gedit, it's better to disallow network access if no plugins require it.

Note that there is also the gedit-plugins git repo, and third-party
gedit plugins are also developed. Those may require network access. The
goal with this commit is to more clearly identify which plugins require
network/internet access, and be able to disable the build of those
plugins, to have a more confined sandbox.