Remove lockdown support
The lockdown support followed the org.gnome.desktop.lockdown GSettings.
But my understanding is that very few GNOME apps implement it, gedit was
a bit the exception.

Removing the lockdown support permits to simplify the gedit code. It
reduces the number of scenarios to test the app. With simpler code,
there will be fewer bugs.

For a GNOME desktop in kiosk mode, I'm not an expert but I think there
are better ways of doing it than using the org.gnome.desktop.lockdown
GSettings. Something for the whole session and all apps at once, without
requiring a detailed implementation in each app.

BTW the implementation was broken in _gedit_window_set_lockdown() with
the GPOINTER_TO_BOOLEAN() in the callback.