Remove GeditAppX11 class
In the main() function, create the parent class instead: GeditApp (this
requires a small change in GeditApp to make it a concrete class instead
of an abstract class).

GeditAppX11 did only one thing (providing the desktop-startup-id in case
the information is needed but not already set). And by looking at the
GTK 3 code, it's now done by GtkApplication (search "desktop-startup-id"
in gtkapplication.c).

This removes some X11-specific code from gedit, which is also a welcome
change since X11 is deprecated and linking against the X11 library
prevents gedit from being a native Wayland app.