Shortcuts window: remove a shortcut that come from gedit-plugins
Ctrl+J comes from the Join/Split Lines plugin that is part of the
gedit-plugins repo (not this one). So when the plugin is not installed
or not activated, Ctrl+J doesn't work.

The Join/Split Lines's keyboard shortcuts are documented in the user
manual of gedit-plugins (in Mallard). There is also Ctrl+Shift+J to
split a line, which was not listed in the Shortcuts window.