gschema: revisit the set of plugins activated by default
Remove time, add sort. And sort the list.

You see, the sort plugin *is* useful :-)

For the time, I don't have a strong opinion, personally I don't think
it's super useful. It's better to keep a small list of plugins activated
by default. It's not a good idea to activate too much stuff by default,
to not make the app and menus "heavy". It's better to keep gedit
lightweight by default. So here we remove one small plugin, and we add
another small one - seems about right.

The time plugin is anyway still *installed* by default and can be
activated easily by going to the preferences.