build: add option to disable installing the externaltools plugin
The externaltools plugin doesn't work on Windows, there is an error when
loading it, it cannot import fcntl (which is specific to Unix).

So, be able to not install the plugin.

Maybe there is a way to make that plugin work on Windows, but I'm far
from an expert in Python, so contributions welcome.