Fix save as behaviour when saving error infobar is showing
As described in #377, attempting to "Save as" with the "unable to save" infobar displayed
(typically caused by trying to save to a directory without the proper permissions) currently
fails silently - the file is not saved, but no error message is given.

This commit causes the Save As function to dismiss the unable to save infobar
(as well as the SAVING_ERROR state), allowing it to behave as expected.

Closes #377