Remove gitlab-ci (CI, Continuous Integration)
The current trend is to "CI" all the things, but in my opinion it slows
down development, you need to babysit the CI pipeline etc, it regularly
fails for reasons that are unrelated to the actual gedit code.

So, in my opinion there are more disadvantages than advantages.

It doesn't mean that automated integration is bad per se, it just needs
to be done differently from my point of view.