filebrowser: remove confirm-trash gsetting, as in Nautilus
Nautilus has removed the gsetting, keeping its default value (true) for
the behavior.

So, always ask for confirmation.

This fixes a crash at startup:
```
GLib-GIO-ERROR: Settings schema 'org.gnome.nautilus.preferences' does
not contain a key named 'confirm-trash'
```

In the future, before using an external GSettings key, we should
introspect at run-time if the *key* exists (not only the GSettings base
object, which was already checked here for NAUTILUS_BASE_SETTINGS with
NAUTILUS_FALLBACK_SETTINGS as a fallback).

We could have kept the confirm-trash gsetting key from
org.gnome.gedit.plugins.filebrowser.nautilus, and still use it, so that
users could configure that aspect, but we think it's better to follow
what Nautilus does, and the fewer options we have, the better (easier to
test).