AppData: new location for the screenshot
Note that data/screenshot.png in this repo needs to be kept for some
time still, because older versions of gedit points to it (especially
because it points to the master branch).

I cannot find an optimal solution unfortunately. On the gnome wiki it's
not ideal, in this repo neither (it makes the main repo bigger to
clone), on a server somewhere neither because gnome doesn't have one for
such purpose.
==> So, create a new gedit-extra repo, similar to gnome-latex-extra.