filebrowser: fix use of deprecated g_pattern function
When briefly looking at the API docs of these two functions, it's hard
to know if they are implemented exactly the same.

But when looking at the *implementation* of g_pattern_match(), it just
calls directly g_pattern_spec_match() with the same args.

Tested with printf-debugging, the pattern correctly matches *.la files
(from a built Autotools project).

So everything should be fine, and the feature still works.