Window: better create open recent menu
Amtk 5.5.2 is needed for
amtk_application_window_create_open_recent_menu_base().

When using amtk_application_window_create_open_recent_menu() (so the
more complete one), it calls g_application_open() which is not suitable
in gedit, because it calls _gedit_cmd_load_files_from_prompt(). From
prompt, if the file doesn't exist, it sets the 'create' param to TRUE,
which creates the file, so it doesn't show a warning infobar if the file
doesn't exist.

When opening a file from the recent menu, we *want* a warning infobar if
the file doesn't exist instead of showing an empty document. (if it
shows an empty document, we may think that the file *exists* and is
empty).