io-error-info-bar: put original error message in details in one case
This parse_gio_error() function needs to be re-read and probably
re-worked a bit, to have more useful / fix error messages displayed in
the infobar.

In this case, it's for G_IO_ERROR_NOT_FOUND and G_IO_ERROR_NOT_DIRECTORY
cases only.

Especially useful for the open recent menu.