tab: after loading, no need to scroll to cursor if already at start
Especially useful if the org.gnome.gedit.preferences.editor
restore-cursor-position gsetting is set to false, and if the file is
loaded without specifying a position.

In that case the cursor will be placed at the start, so no need to setup
the idle+timeout.

(cherry picked from commit be6c7695d457b1054b0616f1609d452f86210923)