About dialog: better list authors and contributors
Add a "Many thanks also to" section.

Remove email addresses for documenters too, there are better means to
communicate with the doc team.

This kind of things is always delicate, it's not nice to remove someone.
But it's nice to thank past contributors, and it's useful to communicate
who are the main authors (number of commits here, one metric needed to
be chosen, and that metric is easily available with git shortlog -sn).

The trailing blank line at the end of authors is to space out the next
section in the credits.

Sort some lists alphabetically, also (by first name, easier).