flatpak: simplify generate-json-manifest.sh
The less the better. A single command goes straight to the point and is
easier to understand, and to modify later on.

If yq is not installed, we get the usual command-not-found error.

When successful, we can then look at the git diff to see if the *.json
file changed.

The set -e is not needed for a single command.

And change a little the yq command, with a link to the project. It
doesn't change the result for the *.json file.

https://gitlab.gnome.org/GNOME/gedit/-/issues/515