tab: don't use gtk_info_bar_set_default_response()
Recently with GTK 3 (since 3.24.14), a simple click on the infobar
activates the default response. It was not intended to be used this way
for gedit. Especially because it's almost unvisible when the mouse is
hover (it's supposed to highlight the infobar differently).

See for example the use of tepl_io_error_info_bar_externally_modified()
and GTK_RESPONSE_OK which was set as the default response. The problem
is that GTK_RESPONSE_OK can mean "Drop changes and reload". So a simple
mis-click drops the unsaved document modifications.

As a result, simplify the code and don't use
gtk_info_bar_set_default_response().