window: make GeditWindowPrivate struct really private
All its external uses have been adapted.

Direct access to struct fields should be used when that's really
necessary (for performance reasons). Otherwise it can be really tricky
to know where in a codebase a certain field is accessed, modified, etc.

Particularly here with such a big number of fields!

Take the occasion also to remove the alignment for the field names. Most
fields were mis-aligned, and it's a burden when modifying code.