statusbar: tab-width-menu: remove auto-indent item
This setting can already be accessed in the Preferences Dialog. No need
to duplicate a lot of settings, otherwise the UI gets clumsy.

Note that the the win.auto-indent action had an effect only on the
current document, while the Preferences Dialog is global to the
application.
For this setting, having only the global config is sufficient, IMHO.

There is also the fact that win.auto-indent doesn't save that setting in
the document metadata.