help: remove scripts
For check_status.sh, its output can be misleading because gedit evolves
continuously and the docs are not always updated accordingly.

For the schematron.sh (which uses gnome-help.sct), I would like to
simplify the *.page files and it would no longer respect all the rules
of the schematron.sh. Basically what I consider sufficient is for the
page to be a valid Mallard file and gives no errors when opening it in
Yelp.