help: remove some <info> sub-elements
For <revision>: git is better at tracking this.

For <credit>: git is also better at displaying such statistics, and
there is also the About dialog to show a list of significant
contributors (including for the docs).

For the <include> legal.xml: IMHO this is not really necessary. Other
smaller projects don't include such information and that's fine.