Window: don't represent window state in the statusbar
It's not really an essential feature.

GeditTab and GtkInfoBars already represent well what's going on in a
GeditWindow. The small icon(s) in the statusbar for the general window
state is thus redundant. When switching to a GeditTab, it's clear enough
what is its state.

gedit_statusbar_set_window_state() will be removed in a further commit.

So, the less code the better (to have fewer chances to trigger bugs).