build: silence one more warning type
There are currently a lot of these warnings, and fixing them all will
take some time. Basic casts can be done, but it's better to see if
changing one of the variable type (gint <-> guint) would be a better
solution. This needs to look at what the code actually does, what
functions are called, etc.