App icons: restore to the previous design
The 'org.gnome.gedit.svg' (deleted by this commit) had several
disadvantages:
- The pencil was way too big.
- The blue color at the bottom was barely visible, especially at small
  sizes.
- No gum? How can you edit what you write with the pencil? (just
  joking ;-).

Later, the GNOME designers have created a new icon for gnome-text-editor
which fixes the above problems. I think the GNOME design team already
knew that there would be a second iteration on the icon... (but that's
just a guess coming from me). No luck for gedit, the new iteration on
the icon went to a new app rather than gedit.

In any case, I prefer the previous design with the old icons. They are
quite well done.