Use TeplSettingsSimpleThemeVariant for the GTK light/dark theme variants
gedit is not fully ready for the 'system' theme variant setting.

The GSettings and the preferences dialog UI for the style/color schemes
(coming from libgedit-gtksourceview) need some work to be able to choose
the one that is set for a light GTK theme, and the one that is set for a
dark GTK theme.

So in the meantime, in order for gedit to be in a releasable state, use
the simplified API from Tepl. Only "force light" and "force dark" are
supported, so gedit doesn't take into account the system setting.

However with some scripting it would be possible to achieve the desired
behavior when the system setting for the GTK theme changes, adapting
automatically the style/color scheme.