WindowTitles: bug fix, rework the update function
- Split set_titles() in 3 functions.
- Use g_get_application_name() in every place where "gedit" was used.
- Use tepl_buffer_get_short_title() to directly get the star if the
  buffer is modified.

About truncation
----------------

Do not care about truncation (remove the related code). GtkHeaderBar
truncates the strings at the end, which is good enough (the window can
be resized in order to see the whole string if needed). The single-title
is also truncated by GTK at the end if there is no GtkHeaderBar.
gnome-shell truncates also the single-title at the end for the window
tooltip in the Overview.

Truncation in the middle with GtkLabel is possible. It would be feasible
to put two custom GtkLabels in the GtkHeaderBar, but in order to do so
(to just adjust where the strings are truncated, keeping all the rest
the same) the implementation of GtkHeaderBar needs to be read and
copied. Relying on implementation details is never a good thing, so it's
better to avoid it. It's also less work to rely on the default behavior
of GtkHeaderBar, of course.

About the bug (now fixed)
-------------------------

When the filename is too long (so it needed to be truncated), the
directory was not shown as subtitle.