MS Windows: add Portable Executable resource icon
Portable Executable, or PE for short:
https://en.wikipedia.org/wiki/Portable_Executable

This associates the icons with the executable (embedding the images
*within* the executable), so that Windows can show the icon in various
situations (in the file explorer, in the taskbar, etc).

The *.ico file was created with this ImageMagick command on Linux:
```
$ cd data/icons/
$ convert \
        org.gnome.gedit-256x256.png \
        org.gnome.gedit-48x48.png \
        org.gnome.gedit-32x32.png \
        org.gnome.gedit-24x24.png \
        org.gnome.gedit-22x22.png \
        org.gnome.gedit-16x16.png \
        gedit.ico
```

The order of the input files should be from the largest size to the
smallest.

An *.ico file was already present in the git repo, but was removed with
commit 23ec52654063d31672b0ec511a826a2ca103a955

However the old *.ico file was bigger in size, and a command-line to
create the file is always better (for reproducibility).

Closes https://gitlab.gnome.org/GNOME/gedit/-/merge_requests/168

Co-authored-by: Sébastien Wilmet <swilmet@mailfence.com>