Make gedit work without patches for MINGW-packages
For MS Windows with MSYS2 and MINGW.

MINGW-packages needed to patch gedit.

It is assumed that get_option('libdir') is the default value of the
install_dir keyword argument for the shared_library() Meson function.
(the patch just removed the install_dir line, to take the default
value).

If the code was useful for other Windows downstream distributors, then
an option can be added to meson_options.txt.